+++ CHANGELOG.md
@@ -11,6 +11,8 @@
         - update help message. [#387](https://github.com/shenwei356/seqkit/issues/387)
     - `seqkit fxtab`:
         - faster alphabet computation (`-a/--alphabet`) with a new data structure. [#388](https://github.com/shenwei356/seqkit/pull/388)
+    - `seqkit translate`:
+        - add options `-s/--out-subseqs` and `-m/--min-len` to write ORFs longer than `x` amino acids as individual records. [#389](https://github.com/shenwei356/seqkit/issues/389)
 - [SeqKit v2.4.0](https://github.com/shenwei356/seqkit/releases/tag/v2.4.0) - 2023-03-17
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/seqkit/v2.4.0/total.svg)](https://github.com/shenwei356/seqkit/releases/tag/v2.4.0)
     - `seqkit`: